pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pptx2marp"
version = "0.1.0"
description = "A Python tool for converting PowerPoint .pptx files to Marp Markdown"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [{ name = "hnrobert" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup",
    "Topic :: Utilities",
]
dependencies = ["python-pptx>=0.6.23"]

[project.urls]
Homepage = "https://github.com/HNRobert/pptx2marp"
Repository = "https://github.com/HNRobert/pptx2marp"

[project.scripts]
pptx2marp = "pptx2marp.cli:main"

[tool.setuptools.packages.find]
include = ["pptx2marp*"]
